Working version of mutating weave
Modified weave() function to either mutate the existing linked list or
create a new linked list. The motivation for this was to explore two
different ways of solving this problem.

Note that the mutating version looks uglier because of all the debug
statements; if they are removed it's cleaner. As the debug statements
probably convey, the mutating version was a bit less intuitive for me to
implement.