CS Question: Is T2 subtree of T1?
To do this question it was necessary to utilize some data structures
provided elsewhere in the repo (i.e. binary tree instantiator and a
stack). To do this it was necessary to reorganize the repo into
submodules as well as add a setup.py so that we can `pip install -e .`
the repo.

Ideally this should have been multiple commits, but given that the
purpose of the re-organization was to facilitate this question it
seems slightly less bad to make this one atomic commit.