Rewrite fibonacci list builder to fix bug
The previous algorithm for building all lists of fibonacci numbers that
sum to a fibonacci number had a bug. Briefly, the previous approach
would build the list of possible lists out of the results for the two
previous fibonacci numbers. However, consider what we would get if we
tried get the lists for the number 8:

    lists for 5: [
        [1, 1, 1, 1, 1],
        [2, 1, 1, 1],
        [2, 2, 1],
        [3, 1, 1],
        [3, 2],
        [5],
    ]

    lists for 3: [
        [1, 1, 1],
        [2, 1],
        [3],
    ]

But 8 can also be made from [2, 2, 2, 2], which we won't generate here.

To resolve this required rebuilding this algorithm to emulate the
create_change algorithm.